componentização de elementos input, informações pessoais,
