--- .idea/modules/app/VitalisApp.app.iml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module version="4">
-  <component name="ModuleClassLoaderOverlays">
-    <paths>
-      <option value="C:\Users\ALOCEL~1\AppData\Local\Temp\overlay6806528276449955890" />
-      <option value="C:\Users\ALOCEL~1\AppData\Local\Temp\overlay2762447854572689664" />
-      <option value="C:\Users\ALOCEL~1\AppData\Local\Temp\overlay14078826646815850977" />
-      <option value="C:\Users\ALOCEL~1\AppData\Local\Temp\overlay15164078201062709462" />
-      <option value="C:\Users\ALOCEL~1\AppData\Local\Temp\overlay2965579588529204006" />
-      <option value="C:\Users\ALOCEL~1\AppData\Local\Temp\overlay687626064930773806" />
-      <option value="C:\Users\ALOCEL~1\AppData\Local\Temp\overlay14435675853019514791" />
-      <option value="C:\Users\ALOCEL~1\AppData\Local\Temp\overlay7938902236773597296" />
-      <option value="C:\Users\ALOCEL~1\AppData\Local\Temp\overlay8734461244782665609" />
-      <option value="C:\Users\ALOCEL~1\AppData\Local\Temp\overlay8953963736397102968" />
-    </paths>
-  </component>
-</module>
