+++ app/src/main/java/com/example/vitalisapp/Activity/Componentes.kt
@@ -1,6 +1,7 @@
 package com.example.vitalisapp.Activity
 
 import android.content.Intent
+import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
@@ -9,6 +10,7 @@ import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.IntrinsicSize
+import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
@@ -19,10 +21,15 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.layout.wrapContentWidth
+import androidx.compose.foundation.lazy.grid.GridCells
+import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
+import androidx.compose.foundation.lazy.grid.items
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.IconButton
@@ -496,6 +503,367 @@ fun CardAtividades( titulo: String) {
                         contentDescription = "seta",
                         modifier = Modifier.size(34.dp)
                     )
+                }
+            }
+        }
+    }
+}
+
+@Composable
+fun InfoCard(title: String, value: String) {
+    Card(
+        shape = RoundedCornerShape(8.dp),
+        modifier = Modifier
+            .width(300.dp)
+            .height(75.dp)
+            .padding(4.dp)
+    ) {
+        Column(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(8.dp),
+            verticalArrangement = Arrangement.Center,
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            Text(
+                text = title,
+                fontSize = 14.sp,
+                fontWeight = FontWeight.Bold,
+                modifier = Modifier.padding(bottom = 4.dp)
+            )
+            Text(
+                text = value,
+                fontSize = 12.sp
+            )
+        }
+    }
+}
+
+@Composable
+fun Tag(text: String) {
+    Box(
+        modifier = Modifier
+            .border(1.dp, Color(72, 183, 90), shape = RoundedCornerShape(16.dp))
+            .background(Color(241, 241, 241), shape = RoundedCornerShape(16.dp))
+            .padding(horizontal = 8.dp, vertical = 4.dp)
+    ) {
+        Text(text = text, color = Color(72, 183, 90), fontSize = 12.sp)
+    }
+}
+
+@Composable
+fun BotaoConcluido() {
+    var isCompleted by remember { mutableStateOf(false) }
+
+    Button(
+        onClick = { isCompleted = !isCompleted },
+        colors = ButtonDefaults.buttonColors(
+            containerColor = if (isCompleted) Color(72, 183, 90) else Color(27, 112, 202)
+        ),
+        shape = RoundedCornerShape(20.dp),
+        modifier = Modifier
+            .padding(16.dp)
+            .fillMaxWidth()
+            .height(50.dp)
+    ) {
+        Image(
+            painter = painterResource(id = R.mipmap.certos),
+            contentDescription = null,
+            modifier = Modifier.size(24.dp)
+        )
+        Spacer(modifier = Modifier.width(8.dp))
+        Text(
+            text = if (isCompleted) "Concluído" else "Marcar como concluído",
+            color = Color.White,
+            fontSize = 16.sp
+        )
+    }
+}
+
+@Composable
+fun PersonalCard(nome: String, especialidade: String, endereco: String) {
+    Card(
+        modifier = Modifier
+            .padding(8.dp)
+            .fillMaxWidth()
+            .shadow(elevation = 4.dp, shape = RoundedCornerShape(8.dp))
+            .border(BorderStroke(1.dp, Color.Gray), shape = RoundedCornerShape(8.dp)),
+        shape = RoundedCornerShape(8.dp),
+        colors = CardDefaults.cardColors(containerColor = Color.White)
+
+    ) {
+        Row(
+            modifier = Modifier.padding(16.dp),
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            Image(
+                painter = painterResource(id = R.mipmap.usuarioperfil),
+                contentDescription = "Profile Picture",
+                modifier = Modifier
+                    .size(60.dp)
+                    .clip(CircleShape),
+                contentScale = ContentScale.Crop
+            )
+            Spacer(modifier = Modifier.width(16.dp))
+            Column {
+                Text(
+                    text = nome,
+                    fontWeight = FontWeight.Bold,
+                    fontSize = 18.sp,
+                    color = Color(72, 183, 90)
+                )
+                Text(
+                    text = especialidade,
+                    fontSize = 14.sp,
+                    color = Color.Black
+                )
+                Text(
+                    text = endereco,
+                    fontSize = 14.sp,
+                    color = Color.Black
+                )
+            }
+        }
+    }
+}
+
+@Composable
+fun UserCard(user: String, meta: String, numero: Int, imagemUSer: Int) {
+    Card(
+        modifier = Modifier.fillMaxWidth(),
+        shape = RoundedCornerShape(12.dp)
+    ) {
+        Row(
+            modifier = Modifier.padding(10.dp),
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            Image(
+                painter = painterResource(id = imagemUSer),
+                contentDescription = "Foto Usuário",
+                modifier = Modifier
+                    .size(65.dp)
+                    .clip(CircleShape),
+            )
+            Spacer(modifier = Modifier.width(20.dp))
+            Column {
+                Text(
+                    text = user,
+                    fontSize = 20.sp,
+                    fontWeight = FontWeight.SemiBold,
+                    color = Color(134, 86, 169)
+                )
+                Row {
+                    Text(text = "Meta:", color = Color(134, 86, 169))
+                    Spacer(modifier = Modifier.width(8.dp))
+                    Text(
+                        text = meta,
+                        fontWeight = FontWeight.Bold,
+                        color = Color(24, 24, 27)
+                    )
+                }
+                Text(
+                    text = "Você tem ${numero} novas mensagens",
+                    fontWeight = FontWeight.Bold,
+                    color = Color.Black
+                )
+            }
+            Spacer(modifier = Modifier.weight(1f))
+
+        }
+    }
+}
+
+@Composable
+fun ExercicioItem(exercicios: List<String>) {
+    val contexto = LocalContext.current
+
+    LazyVerticalGrid(
+        columns = GridCells.Fixed(2),
+        contentPadding = PaddingValues(top = 20.dp),
+        horizontalArrangement = Arrangement.spacedBy(16.dp),
+        verticalArrangement = Arrangement.spacedBy(16.dp)
+    ) {
+        items(exercicios) { exercicio ->
+            CardReceita(exercicio) {
+                val detalheExercicio = Intent(contexto, DetalheExercicio::class.java)
+                contexto.startActivity(detalheExercicio)
+            }
+        }
+    }
+}
+
+@Composable
+fun GridReceita(receitas: List<String>) {
+    val contexto = LocalContext.current
+
+    LazyVerticalGrid(
+        columns = GridCells.Fixed(2),
+        contentPadding = PaddingValues(top = 20.dp),
+        horizontalArrangement = Arrangement.spacedBy(16.dp),
+        verticalArrangement = Arrangement.spacedBy(16.dp)
+    ) {
+        items(receitas) { receita ->
+            CardReceita(receita){
+                val receita = Intent(contexto, Receita::class.java)
+                contexto.startActivity(receita)
+            }
+        }
+    }
+}
+
+@Composable
+fun BarraPesquisa(modifier: Modifier = Modifier,  onTextChange: (String) -> Unit) {
+    var searchText by remember { mutableStateOf("") }
+
+    OutlinedTextField(
+        value = searchText,
+        onValueChange = { searchText = it
+            onTextChange(it)},
+        modifier = modifier
+            .clip(RoundedCornerShape(30.dp))
+            .background(Color(255, 255, 255))
+            .border(2.dp, Color(0, 0, 0), RoundedCornerShape(30.dp)),
+        placeholder = { Text(stringResource(R.string.pesquisa),
+            fontFamily = MavenPro,) },
+        leadingIcon = {
+            Image(
+                painter = painterResource(id = R.mipmap.lupa),
+                contentDescription = "Lupa",
+                modifier = Modifier.size(25.dp)
+            )
+        }
+    )
+}
+
+@Composable
+fun DateInput(modifier: Modifier = Modifier,  onTextChange: (String) -> Unit) {
+    var searchDate by remember { mutableStateOf("") }
+
+    Column(modifier = modifier) {
+        Text(
+            text = "Data",
+            color = Color(0xFF1A1A1A),
+            fontFamily = MavenPro,
+            fontSize = 12.sp,
+            fontWeight = FontWeight.Bold,
+            letterSpacing = 0.6.sp,
+            modifier = Modifier.padding(start = 16.dp)
+        )
+        OutlinedTextField(
+            value = searchDate,
+            onValueChange = {searchDate = it
+                onTextChange(it)},
+            modifier = Modifier.fillMaxWidth(),
+            placeholder = { Text("Data",fontFamily = MavenPro) },
+            leadingIcon = {
+                Image(
+                    painter = painterResource(id = R.mipmap.calendario),
+                    contentDescription = "Calendário",
+                    modifier = Modifier.size(25.dp)
+                )
+            },
+            shape = RoundedCornerShape(30.dp)
+        )
+    }
+}
+
+@Composable
+fun ItemInfo(label: String, valor: String, modifier: Modifier = Modifier) {
+    Column(modifier = modifier) {
+        Text(
+            text = label,
+            fontFamily = MavenPro,
+            fontSize = 18.sp,
+            fontWeight = FontWeight.Bold,
+            color = Color.Black
+        )
+        Text(
+            text = valor,
+            fontFamily = MavenPro,
+            fontSize = 16.sp,
+            color = Color(24, 24, 27),
+            modifier = Modifier.padding(top = 8.dp)
+        )
+    }
+}
+
+@Composable
+fun CartaoInfo(
+    tipoUsuario: String,
+    imagemUsuario: Int,
+    nome: String,
+    email: String,
+    emailRecuperacao: String = "",
+    nickname: String,
+    aniversario: String,
+    sexo: String,
+    especialidade: String = "",
+    graduacao: String = ""
+) {
+
+    val corTexto = if (tipoUsuario == "personal") Color(134, 86, 169) else Color(72, 183, 90)
+    val corBotao = if (tipoUsuario == "personal") R.mipmap.botaopersonal else R.mipmap.botaomural
+
+    Card(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(top = 10.dp, start = 16.dp, end = 16.dp),
+        shape = RoundedCornerShape(24.dp),
+        elevation = CardDefaults.cardElevation(defaultElevation = 4.dp),
+        colors = CardDefaults.cardColors(containerColor = Color.White)
+    ) {
+        Column(
+            modifier = Modifier.padding(24.dp),
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            Text(
+                text = stringResource(R.string.info_pessoal),
+                fontFamily = MavenPro,
+                fontSize = 24.sp,
+                fontWeight = FontWeight.SemiBold,
+                color = corTexto,
+                modifier = Modifier.padding(bottom = 16.dp)
+            )
+            Box(
+                contentAlignment = Alignment.Center
+            ) {
+                Image(
+                    painter = painterResource(id = imagemUsuario),
+                    contentDescription = "Foto Usuário",
+                    modifier = Modifier
+                        .size(150.dp)
+                        .clip(CircleShape)
+                        .border(2.dp, Color.Gray, CircleShape)
+                )
+                Image(
+                    painter = painterResource(id = corBotao),
+                    contentDescription = "Botão",
+                    modifier = Modifier
+                        .size(50.dp)
+                        .clip(CircleShape)
+                        .clickable { onEditClick() }
+                        .align(Alignment.BottomEnd)
+                )
+            }
+            Column(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(top = 32.dp)
+            ) {
+                ItemInfo("Nome completo:", nome)
+                ItemInfo("E-mail principal:", email)
+                ItemInfo("Nickname:", nickname)
+                ItemInfo("Data de nascimento:", aniversario)
+                ItemInfo("Sexo:", sexo)
+
+
+                if (tipoUsuario == "personal") {
+                    ItemInfo("Especialidade:", especialidade)
+                    ItemInfo("Data de formação:", graduacao)
+                }
+
+                if (tipoUsuario == "usuario") {
+                    ItemInfo("E-mail de recuperação:", emailRecuperacao)
                 }
             }
         }
